structure normalization parallelized

+++ appsite/etl/management/commands/register.py
@@ -2,6 +2,7 @@
 import logging
 from typing import List
 
+from django.conf import settings
 from django.core.management.base import BaseCommand
 
 from etl.tasks import *
@@ -19,11 +20,19 @@ class Command(BaseCommand):
         _register()
 
 
+def register_file_records(structure_file_id: int):
+    task_list = \
+        (count_and_save_file_task.s(structure_file_id) |
+         register_file_record_chunk_mapper.s(register_file_record_chunk_task.s()))
+    return task_list.delay()
+
+
 def _register():
 
-    for f in StructureFile.objects.all():
-        logger.info("deleting %s", f)
-        f.delete()
+    if settings.INIT_SYSTEM:
+        for f in StructureFile.objects.all():
+            logger.info("deleting %s", f)
+            f.delete()
 
     #file_collections = FileCollection.objects.all()
     file_collections = FileCollection.objects.filter(id=4)
